Version 0.3.20
Version 0.3.20
Added an AutoCensor system. The system will automatically censor words
on the banned list. For example, if the word "gris" was on the banned
list, it would be automatically censored to "g***"
Added On-Demand Bug Report to navigation drawer. It will allow you to
send a bug report whenever you need to, even when the bug catcher
doesn't pop up
Added a link on the top of boards that are part of a split list to go to
the split list
Modified account storage encryption to use a randomly generated salt,
unique to each device, previously stored accounts will be automatically
converted to the new encryption system