re-theme action bar
action bar background is completely transparent as far as the
overarching theme is concerned, and this is what ptr looks at. each
activity then sets it's action bar's activity to the appropriate (light
or dark) transparent theme drawable with a color filter applied to
change the holo blue "separator" line in the drawable to whatever the
accent color is. since ptr uses the fully transparent action bar from
the theme, this "separator" "disappears" when the ptr is in a loading
state, then "reappears" once finished